- Refactor ClientAppAdapter
