+++ modules/apps/src/main/java/com/axway/apim/appimport/ClientApplicationImportApp.java
@@ -1,7 +1,7 @@
 package com.axway.apim.appimport;
 
+import com.axway.apim.ClientAppAdapter;
 import com.axway.apim.adapter.APIManagerAdapter;
-import com.axway.apim.adapter.client.apps.ClientAppAdapter;
 import com.axway.apim.adapter.client.apps.ClientAppFilter;
 import com.axway.apim.api.model.apps.ClientApplication;
 import com.axway.apim.appimport.adapter.ClientAppConfigAdapter;
